Final Backend Push v5

+++ backend/db/models/venue.js
@@ -35,11 +35,11 @@ module.exports = (sequelize, DataTypes) => {
       allowNull: false
     },
     lat: {
-      type: DataTypes.DECIMAL(10, 7),
+      type: DataTypes.FLOAT(10, 7),
       allowNull: true
     },
     lng: {
-      type: DataTypes.DECIMAL(10, 7),
+      type: DataTypes.FLOAT(10, 7),
       allowNull: true
     }
   }, {
